requirements.txt
anndata==0.9.2
h5py==3.7.0
matplotlib==3.7.1
numpy==1.22.3
pandas==2.0.3
scanpy==1.9.5
scipy==1.9.1
seaborn==0.12.2
scglue==0.3.2
lisa2[coverage]
tables
tqdm
ipywidgets